laudos_api/requirements.txt
Django==2.2.*
django-filter==2.2.*
django-dotenv==1.4.*
django-cors-headers==3.1.*
django-extensions==2.2.*
django-crispy-forms==1.7.*
djangorestframework==3.10.*
drf-yasg==1.16.*
coreapi==2.3.*
markdown==3.1.*
packaging
smart-getenv
coverage
ipython
jupyter
black
pep8
autopep8
pylint
pylint-django